pyproject.toml
[project]
name = "cli-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic[bedrock]>=0.96.0",
    "fastapi>=0.136.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.45.0",
]
